% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{flatness_indices.default}
\alias{flatness_indices.default}
\title{Default method for S3 generic function \code{flatness_indices}}
\usage{
\method{flatness_indices}{default}(rkhists, ...)
}
\arguments{
\item{rkhists}{an object containing (rank) histograms}

\item{...}{other arguments}
}
\value{
No return value, called for side effects (generate an error message).
}
\description{
Just generate an error.
}
