% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{flatness_indices.matrix}
\alias{flatness_indices.matrix}
\title{Method for S3 generic function \code{flatness_indices} and objects with class
\code{matrix}}
\usage{
\method{flatness_indices}{matrix}(rkhists, col_wise = FALSE, ...)
}
\arguments{
\item{rkhists}{an object containing (rank) histograms}

\item{col_wise}{a logical. Are the rank histograms stored column wise, that
is one rank histogram in each column?}

\item{...}{other arguments}
}
\value{
See \emph{Value} in function \code{flatness_indices}
}
\description{
This function is the method called when using the S3 generic function
\code{flatness_indices} with an object of class \code{matrix}.
}
\details{
The input matrices is supposed to contain one rank histograms in each row,
unless \code{col_wise == TRUE}.
}
