% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_orthonormal}
\alias{is_orthonormal}
\title{Check whether a vector set is orthonormal}
\usage{
is_orthonormal(set, tol = 1e-04)
}
\arguments{
\item{set}{a matrix. The convention used here is that each row of \code{set}
contains a vector.}

\item{tol}{a positive numeric. The accepted tolerance for the conditions of
orthonormality.}
}
\value{
TRUE if the set of vectors is orthonormal, FALSE otherwise.
}
\description{
Check whether a vector set is orthonormal
}
