% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviates.R
\name{make_JP_ready}
\alias{make_JP_ready}
\title{Modify a set of \code{rkhist} objects so that they can be used in the
Jolliffe-Primo flatness test}
\usage{
make_JP_ready(x, verbose = TRUE)
}
\arguments{
\item{x}{an \code{rkhist} object.}

\item{verbose}{a logical. Should the result of the check be displayed?}
}
\value{
An \code{rkhist} object, containing the modified vectors set obeying
the requirements for the Jolliffe-Primo flatness test.
}
\description{
The Jolliffe-Primo flatness test requires deviate vectors that are
orthonormal and that each has components summing to zero. This function
ensures this by using, if it is required, the Grahm-Schmidt method to make
the set orthonormal and also makes the sum of each vector's components equal
to zero.
}
\details{
Note this procedure may greatly change the shape of vectors.
}
