% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPtest.R
\name{print.JPtest}
\alias{print.JPtest}
\title{Print method for a \code{JPtest} object}
\usage{
\method{print}{JPtest}(x, what = c("projections", "statistics", "pvalues"), ...)
}
\arguments{
\item{x}{the \code{JPtest} object that must be printed.}

\item{what}{a character vector. What must be printed? Can contain any or
several among c("projections", "statistics", "pvalues").}

\item{...}{further arguments. Passed to \code{print} method for
matrix objects.}
}
\value{
No return value, called for side effects.
}
\description{
Print method for a \code{JPtest} object
}
