% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkhist.R
\name{print.rkhist}
\alias{print.rkhist}
\title{Print an S3 object of class \code{rkhist}}
\usage{
\method{print}{rkhist}(x, what = "counts", ...)
}
\arguments{
\item{x}{an \code{rkhist} S3 object containing the rank histogram to
print. See function \code{flatness::rkhist}.}

\item{what}{a character string taking its value in
\code{c("counts", "percents", "proportions")}. What must be printed.}

\item{...}{further arguments, passed to \code{print}.}
}
\value{
No return value, called for side effects.
}
\description{
Print an S3 object of class \code{rkhist}, using the same layout as
\code{base::print.table}.
}
