% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkhist.R
\name{rbind_rkhists}
\alias{rbind_rkhists}
\title{Stack objects of class \code{rkhist}}
\usage{
rbind_rkhists(..., names = NULL)
}
\arguments{
\item{...}{matrix, rkhist objects or vectors to bind by row.}

\item{names}{a character vector of names to give to each rank histogram
(i.e. each row of the resulting matrix). Although its length must be equal to
the number of row in the resulting matrix, \code{NA}s are allowed.}
}
\value{
An object of S3 class \code{rkhist}
}
\description{
Use \code{rbind} to stack several rank histograms (stored in S3 objects of
class \code{rkhist}, or matrix, or vectors).
}
