% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkhist.R
\name{rkhist.default}
\alias{rkhist.default}
\title{Default method for S3 generic function \code{rkhist}}
\usage{
\method{rkhist}{default}(fcst, obs, ...)
}
\arguments{
\item{fcst}{an object containing the ensemble forecasts.}

\item{obs}{an object containing the observation associated to the forecast in
\code{fcst}.}

\item{...}{additional arguments.}
}
\value{
No return value, called for side effects (generate an error message).
}
\description{
Just generate an error.
}
